services/head/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pyyaml>=6.0
fastmcp>=0.4
